changed itemId to be a Long

+++ Book.java
@@ -1,5 +1,5 @@
 public class Book extends Item{
-    Book(String section, String title, String author, String publisher, int itemID) {
+    Book(String section, String title, String author, String publisher, long itemID) {
         setSection(section);
         setTitle(title);
         setAuthor(author);
