+++ Newspaper.java
@@ -1,5 +1,5 @@
 public class Newspaper extends Item{
-    Newspaper(String section, String title, String author, String publisher, int itemID){
+    Newspaper(String section, String title, String author, String publisher, long itemID){
         setSection(section);
         setTitle(title);
         setAuthor(author);
